refactor(react-server): organize features
